# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Torneo de Tenis API",
        "contact": {
            "email": "julian.andres.arguello@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/profiles/{id}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Obtener los detalles de un jugador",
                "description": "Devuelve los detalles de un jugador por ID, incluyendo información adicional si es masculino o femenino",
                "operationId": "23416b602d56220247b2f74592698147",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jugador encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la operación",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Mensaje de la operación",
                                            "type": "string",
                                            "example": "ok"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MalePlayerData"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jugador no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Estado de la operación",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string",
                                            "example": "No existe un Jugador con el ID proporcionado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MalePlayerData": {
                "description": "Datos de un jugador masculino",
                "properties": {
                    "id": {
                        "description": "ID único del jugador",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nombre del jugador",
                        "type": "string",
                        "example": "Julián Argüello"
                    },
                    "skill_level": {
                        "description": "Nivel de habilidad del jugador",
                        "type": "integer",
                        "example": 92
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T16:31:13.000000Z"
                    },
                    "deleted_at": {
                        "description": "Fecha de eliminación",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "male_player": {
                        "description": "Datos específicos del jugador masculino",
                        "properties": {
                            "player_id": {
                                "description": "ID del jugador asociado",
                                "type": "integer",
                                "example": 1
                            },
                            "stength": {
                                "description": "Fuerza del jugador",
                                "type": "integer",
                                "example": 228
                            },
                            "movement_speed": {
                                "description": "Velocidad de movimiento del jugador",
                                "type": "number",
                                "format": "float",
                                "example": 9.34
                            },
                            "created_at": {
                                "description": "Fecha de creación",
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T16:31:13.000000Z"
                            },
                            "updated_at": {
                                "description": "Fecha de actualización",
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01T16:31:13.000000Z"
                            },
                            "deleted_at": {
                                "description": "Fecha de eliminación",
                                "type": "string",
                                "format": "date-time",
                                "example": null,
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}